+++ package.json
@@ -20,6 +20,7 @@
     "start": "browser-sync start --server --no-ui --files \"js/**/*.js\"",
     "lint": "eslint js/"
   },
+  "type": "module",
   "engines": {
     "node": "16.x",
     "npm": "8.x"
